# docs/knowledge-base/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AppointmentOut": {
        "description": "An appointment as returned to clients.",
        "properties": {
          "cancelled_late": {
            "title": "Cancelled Late",
            "type": "boolean"
          },
          "doctor_id": {
            "title": "Doctor Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "slot_id": {
            "title": "Slot Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          }
        },
        "required": [
          "id",
          "patient_id",
          "doctor_id",
          "slot_id",
          "status",
          "reason",
          "cancelled_late"
        ],
        "title": "AppointmentOut",
        "type": "object"
      },
      "AppointmentStatus": {
        "description": "The states an appointment can occupy. Stored as its string value.",
        "enum": [
          "requested",
          "confirmed",
          "checked_in",
          "in_progress",
          "completed",
          "cancelled",
          "no_show"
        ],
        "title": "AppointmentStatus",
        "type": "string"
      },
      "BookingRequest": {
        "description": "A patient's request to book an open slot (story B2).",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason for visit.",
            "title": "Reason"
          },
          "slot_id": {
            "description": "The open slot to book.",
            "title": "Slot Id",
            "type": "integer"
          }
        },
        "required": [
          "slot_id"
        ],
        "title": "BookingRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Credentials for obtaining a token pair (story A3).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password to verify.",
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "Payload to exchange a valid refresh token for a new token pair (story A4).",
        "properties": {
          "refresh_token": {
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "description": "Payload for self-service patient registration (story A1).\n\nOnly patients self-register; staff accounts are admin-provisioned (A2), so\nthis request carries no role field \u2014 the service assigns the PATIENT role.",
        "properties": {
          "email": {
            "description": "Login email; must be unique.",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name.",
            "title": "Full Name"
          },
          "password": {
            "description": "Plaintext password (min 8 chars); hashed server-side.",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "Role": {
        "description": "A user's role. Inherits from ``str`` so it serializes as its plain value\n(e.g. ``\"patient\"``) in JSON, tokens, and the database without extra mapping.",
        "enum": [
          "patient",
          "nurse",
          "doctor",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
      "SlotCreate": {
        "description": "A doctor's request to publish one availability slot (story B1).",
        "properties": {
          "end_at": {
            "description": "Slot end (timezone-aware); must be after start.",
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "start_at": {
            "description": "Slot start (timezone-aware).",
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "SlotCreate",
        "type": "object"
      },
      "SlotOut": {
        "description": "A published availability slot.",
        "properties": {
          "doctor_id": {
            "title": "Doctor Id",
            "type": "integer"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_booked": {
            "title": "Is Booked",
            "type": "boolean"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "doctor_id",
          "start_at",
          "end_at",
          "is_booked"
        ],
        "title": "SlotOut",
        "type": "object"
      },
      "TokenPair": {
        "description": "The access + refresh tokens returned on login/refresh (stories A3, A4).\n\n``token_type`` is the OAuth2 bearer convention so standard clients (and the\nSwagger UI) know how to send the access token: ``Authorization: Bearer ...``.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "Transition": {
        "description": "The named actions that move an appointment between states.",
        "enum": [
          "confirm",
          "check_in",
          "begin",
          "complete",
          "cancel",
          "no_show",
          "reschedule"
        ],
        "title": "Transition",
        "type": "string"
      },
...
      "UserOut": {
        "description": "Public representation of a user account (never includes the password hash).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first medical portal. JSON API for accounts, appointments, clinical records, and prescriptions. The browser UI is server-rendered (Jinja2 + HTMX) by this same application.",
    "title": "HealthyVytals API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/appointments": {
      "post": {
        "description": "Book an available slot for the authenticated patient (stories B2, B3).\n\nA taken slot or a buffer conflict raises ``SlotConflict`` \u2192 409; an unknown\nslot is a 404.",
        "operationId": "book_appointment_api_v1_appointments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Book an open slot (patient only)",
        "tags": [
          "appointments"
        ]
      }
    },
...
    "/api/v1/appointments/slots": {
      "post": {
        "description": "Publish a bookable slot for the authenticated doctor (story B1).",
        "operationId": "publish_slot_api_v1_appointments_slots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish an availability slot (doctor only)",
        "tags": [
          "appointments"
        ]
      }
    },
...
    "/api/v1/appointments/slots/open/{doctor_id}": {
      "get": {
        "description": "Return a doctor's unbooked slots so a patient can choose one (story B2).",
        "operationId": "list_open_slots_api_v1_appointments_slots_open__doctor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doctor_id",
            "required": true,
            "schema": {
              "title": "Doctor Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SlotOut"
                  },
                  "title": "Response List Open Slots Api V1 Appointments Slots Open  Doctor Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a doctor's open slots (any authenticated user)",
        "tags": [
          "appointments"
        ]
      }
    },
...
    "/api/v1/appointments/{appointment_id}/transitions/{transition}": {
      "post": {
        "description": "Apply a state-machine transition (stories B4, B6; \u00a75.1).\n\nAny authenticated user may call this; the pure state machine decides whether\nthe caller's role may make the requested move (patient cancel/reschedule,\nnurse check-in/no-show, doctor confirm/begin/complete), returning 409 for an\nillegal move and 403 when a patient targets someone else's appointment.\n``transition`` is validated against the Transition enum by FastAPI (422 if\nunknown).",
        "operationId": "transition_appointment_api_v1_appointments__appointment_id__transitions__transition__post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "transition",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Transition"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance an appointment through its lifecycle",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Verify credentials and return an access + refresh token pair (story A3).\n\nInvalid credentials raise ``InvalidCredentials`` \u2192 401, with an identical\nerror whether the email is unknown or the password is wrong.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log in and obtain tokens",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Exchange a valid refresh token for a fresh token pair (story A4).\n\nAn access token presented here is rejected \u2014 only refresh tokens are accepted.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh the token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Self-service patient registration (story A1).\n\nReturns the created account as :class:`UserOut` (never the password hash).\nA duplicate email raises ``EmailAlreadyRegistered`` \u2192 409 via the typed-error\nhandlers.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a new patient account",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}